Add loopback address to director_ssl certificate
Seems that with an external_ip enabled on the bosh director the
health_monitor uses the internal loopback address when attempting to
health check the director. Since this address isn't a listed address on
the director_ssl.certificate the health_monitor fails to health check
the director. Perhaps some misconfiguration has occurred when using an
external_ip?