Delete BOSH environment when removed from the list
Using the requested students list (students.auto.tfvars.json), and the
provisioned students list (bosh director state files) we can generate a
list of environments to remove. Differenciated between both lists we
check whether a state file within the director-states S3 bucket
directory exists, if it does, the BOSH Environment should be deleted.
If the information required to remove the BOSH Environment is no longer
available within the terraform state then we remove the creds and state
files for the differenciated students.